new feature:added sidebar
