+++ src/App.js
@@ -1,17 +1,21 @@
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  Switch,
+} from "react-router-dom";
 import LandingPage from "./pages/user/landingPage";
 import MuiNavBar from "./component/navigation/users";
 import ProductList from "./pages/user/productList";
 import ProductListCategory from "./pages/user/productList/category";
 import ProductListSearch from "./pages/user/productList/search";
-import ProductCard from "./component/productCard";
 import ProductDetail from "./pages/user/productDetail";
-import Test from "./pages/admin/warehouseAdmin/inputProducts/index";
-
+import Test from "./pages/admin/warehouseAdmin/modifyProducts";
 function App() {
   return (
     <Router>
       {/* <MuiNavBar /> */}
+
       <Routes>
         <Route path="/" element={<LandingPage />} />
         <Route path="/product-list" element={<ProductList />} />
