--- src/pages/admin/warehouseAdmin/inputProducts/style.css
@@ -1,7 +1,0 @@
-.Selector {
-  width: 60%;
-  margin-left: 25%;
-}
-.Editor {
-  width: 70%;
-}
