+++ src/pages/admin/warehouseAdmin/modifyProducts/modal/editProduct/addVariant/addVariant/style.css
